requirements.txt
aiohttp==3.8.5 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
amqp==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
billiard==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
celery==5.3.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==41.0.2 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
fastapi-cache2==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.82.0 ; python_version >= "3.10" and python_version < "4.0"
fcd-torch==1.0.7 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.12.0 ; python_version >= "3.10" and python_version < "4.0"
flower==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
gevent==22.10.2 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and python_version < "4.0" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64" or platform_python_implementation == "CPython")
gunicorn[standard]==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
humanize==4.6.0 ; python_version >= "3.10" and python_version < "4.0"
identify==2.5.24 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
kombu==5.3.1 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
minio==7.1.15 ; python_version >= "3.10" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.11" and python_version < "4.0"
networkx==3.1 ; python_version >= "3.10" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.3 ; python_version >= "3.10" and python_version < "4.0"
openbabel==3.1.1.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
pendulum==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.5.1 ; python_version >= "3.10" and python_version < "4.0"
pre-commit==2.21.0 ; python_version >= "3.10" and python_version < "4.0"
prometheus-client==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.38 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.6 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.8 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-environ==0.4.54 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pytzdata==2020.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
rdkit==2023.3.1 ; python_version >= "3.10" and python_version < "4.0"
redis==4.5.5 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.9.3 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.8.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.15 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.19.1 ; python_version >= "3.10" and python_version < "4.0"
sympy==1.12 ; python_version >= "3.10" and python_version < "4.0"
torch==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
torchvision==0.15.2 ; python_version >= "3.10" and python_version < "4.0"
tornado==6.3.2 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.6.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.18.3 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.18.3 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
vine==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
virtualenv==20.23.0 ; python_version >= "3.10" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.11" and python_version < "4.0"
zope-event==4.6 ; python_version >= "3.10" and python_version < "4.0"
zope-interface==6.0 ; python_version >= "3.10" and python_version < "4.0"
